+++ src/algorithms/brute_force.c
@@ -2,8 +2,8 @@
 #include "UI/TUI_func.h"
 #include "data_operations/route_struct.h"
 
-static inline void swap_numbers(uint8_t* a, uint8_t* b){
-    uint8_t temp = *a;
+static inline void swap_numbers(uint16_t* a, uint16_t* b){
+    uint16_t temp = *a;
     *a = *b;
     *b = temp;
 }
@@ -21,14 +21,14 @@ Route* brute_force(uint32_t* distances, size_t num_points){
         return NULL;
     }
     
-    result->city_order = malloc(num_points * sizeof(uint8_t));
+    result->city_order = malloc(num_points * sizeof(uint16_t));
     if(!result->city_order){
         print_error("Result->city_order alloc failed in BF algorithm, exiting...\n");
         free(result);
         return NULL;
     }
 
-    uint8_t* temp_permutation = malloc(num_points * sizeof(uint8_t));
+    uint16_t* temp_permutation = malloc(num_points * sizeof(uint16_t));
     if(!temp_permutation){
         print_error("Temp_permutation alloc failed in BF algorithm, exiting...\n");
         free(result->city_order);
@@ -53,7 +53,7 @@ Route* brute_force(uint32_t* distances, size_t num_points){
         if(temp_dist < result->distance_u){
 
             result->distance_u = temp_dist;
-            memcpy(result->city_order, temp_permutation, num_points * sizeof(uint8_t));
+            memcpy(result->city_order, temp_permutation, num_points * sizeof(uint16_t));
 
         }
